requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0

# Audio processing
librosa>=0.9.0
webrtcvad>=2.0.10

# Google Docs API
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Anki package generation
genanki>=0.13.0

# Forced alignment
montreal-forced-alignment>=2.0.0
phonemizer>=3.0.0

# Testing framework
pytest>=7.0.0
hypothesis>=6.0.0

# Development dependencies
black>=22.0.0
flake8>=4.0.0
mypy>=0.950